move fixtures from common app into conftest.py, so they can be available in all tests
